Added an Explorer entity. It can be used for taking screenshots in game, and exploring maps.
-Changed scripts:
	-Camera.py
	-Explorer.py
	-LocalGame.py
	-CommandsUser.py
	-GameControl (__init__.py)

As well as prog to include the new entity.

I believe it was programmed correctly, and spawning a local player will delete it. I would appreciate it if someone checked the changes over, and let me know if I did it properly, as this was mainly to help me learn the entities system.

Next step: Crouching improvements.