Hi. I've been working on stuff. New stuff: 	• Working prototype for Inventory System. 	• Working prototype for Gun Simulation (user control, animation system)
IMPORTANT WARNING!: YOU MUST USE /defaultOptions() IN THE TERMINAL FOR THE GAME TO WORK PROPERLY (or just delete FPS_options.txt :)

There is currently no graphical representation of the gun simulation. You currently have to check the terminal to see what's going on with your gun.

Controls (defaults):
	R = Auto-Reload
	LMB = Shoot
	T = Bolt Release
	F = Cock

Currently, R is a simplified reload, so the Bolt Release and Cock controls are not needed at all to use the gun. When you start in the scene, your gun does not have a magazine in it, so you'll have to do an initial reload before you can start shooting.

Also, none of the controls (except for shooting) works while the terminal is up, even though you need to check the terminal to see what's going on.

I'll work on cheesey graphical assets, and the ballistics system later.
	-Chase